+++ sdk/cosmos/azure-cosmos/azure/cosmos/_base.py
@@ -30,6 +30,7 @@ import binascii
 from typing import Dict, Any
 
 from urllib.parse import quote as urllib_quote
+from urllib.parse import urlsplit
 
 from azure.core import MatchConditions
 
@@ -661,6 +662,11 @@ def ParsePaths(paths):
             tokens.append(token)
 
     return tokens
+
+
+def create_scope_from_url(url):
+    parsed_url = urlsplit(url)
+    return parsed_url.scheme + "://" + parsed_url.hostname + "/.default"
 
 
 def validate_cache_staleness_value(max_integrated_cache_staleness):